+++ src/app/app-routing.module.ts
@@ -5,26 +5,18 @@ import { authGuard } from './guard/auth/auth.guard';
 
 const routes: Routes = [
   {
-    path: Path.Root,
-    children: [
-      {
-        path: Path.Login,
-        loadChildren: () =>
-          import('./login/login.module').then((m) => m.LoginModule),
-      },
-
-      {
-        path: Path.Resource,
-        loadChildren: () =>
-          import('./resources/resources.module').then((m) => m.ResourcesModule),
-
-        canMatch: [authGuard],
-      },
-
-      { path: '**', redirectTo: Path.Resource },
-    ],
+    path: Path.Login,
+    loadChildren: () =>
+      import('./login/login.module').then((m) => m.LoginModule),
   },
-  { path: '**', redirectTo: Path.Root },
+  {
+    path: Path.Resource,
+    loadChildren: () =>
+      import('./resources/resources.module').then((m) => m.ResourcesModule),
+    canMatch: [authGuard],
+  },
+
+  { path: '**', redirectTo: Path.Resource },
 ];
 
 @NgModule({