+++ src/app/enum/path.ts
@@ -1,5 +1,4 @@
 export enum Path {
-  Root = 'horo-admin',
   Login = 'login',
   Resource = 'resources',
 }