+++ src/app/guard/auth/auth.guard.ts
@@ -7,5 +7,5 @@ export const authGuard: CanMatchFn = (route, segments) => {
   const user = inject(AuthService);
   const router = inject(Router);
   if (user.isAuth) return true;
-  else return router.parseUrl(`${Path.Root}/${Path.Login}`);
+  else return router.parseUrl(`${Path.Login}`);
 };