+++ src/app/resources/resources.module.ts
@@ -1,7 +1,9 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { FormsModule } from '@angular/forms';
-import { NgbModalModule, NgbDropdownModule, NgbNav, NgbNavLinkButton, NgbNavItem, NgbTooltipModule } from '@ng-bootstrap/ng-bootstrap';
+import { 
+  // NgbModalModule, NgbDropdownModule, NgbNav, NgbNavLinkButton, NgbNavItem, 
+  NgbTooltipModule } from '@ng-bootstrap/ng-bootstrap';
 
 import { ResourcesRoutingModule } from './resources-routing.module';
 import { StorageCommonModule } from '../common/storage-common.module';
@@ -22,11 +24,11 @@ import { NativeFormComponent } from './natives/native-form/native-form.component
     ResourcesRoutingModule,
     FormsModule,
     StorageCommonModule,
-    NgbModalModule,     // 新增 NgbModalModule
-    NgbDropdownModule,  // 保留原有组件
-    NgbNav,
-    NgbNavLinkButton,
-    NgbNavItem,
+    // NgbModalModule,     // 新增 NgbModalModule
+    // NgbDropdownModule,  // 保留原有组件
+    // NgbNav,
+    // NgbNavLinkButton,
+    // NgbNavItem,
     NgbTooltipModule,
   ],
 })